+++ packages/vendure-plugin-accept-blue/src/accept-blue-plugin.ts
@@ -1,14 +1,17 @@
 import { PluginCommonModule, VendurePlugin } from '@vendure/core';
-import { DefaultSubscriptionStrategy } from '.';
 import { SubscriptionStrategy } from '../../util/src/subscription/subscription-strategy';
 import { AcceptBlueService } from './api/accept-blue-service';
 import { acceptBluePaymentHandler } from './api/accept-blue-handler';
 import { PLUGIN_INIT_OPTIONS } from './constants';
 import { commonApiExtensions } from './api/api-extensions';
 import { AcceptBlueCommonResolver } from './api/accept-blue-common-resolvers';
+import { AcceptBlueController } from './api/accept-blue-controller';
+import { DefaultSubscriptionStrategy } from '../../util/src/subscription/default-subscription-strategy';
+import { rawBodyMiddleware } from '../../util/src/raw-body.middleware';
 
 interface AcceptBluePluginOptionsInput {
   subscriptionStrategy?: SubscriptionStrategy;
+  vendureHost: string;
 }
 
 export type AcceptBluePluginOptions = Required<AcceptBluePluginOptionsInput>;
@@ -23,6 +26,7 @@ export type AcceptBluePluginOptions = Required<AcceptBluePluginOptionsInput>;
     schema: commonApiExtensions,
     resolvers: [AcceptBlueCommonResolver],
   },
+  controllers: [AcceptBlueController],
   providers: [
     AcceptBlueService,
     {
@@ -32,6 +36,11 @@ export type AcceptBluePluginOptions = Required<AcceptBluePluginOptionsInput>;
   ],
   configuration: (config) => {
     config.paymentOptions.paymentMethodHandlers.push(acceptBluePaymentHandler);
+    config.apiOptions.middleware.push({
+      route: '/accept-blue/*',
+      handler: rawBodyMiddleware,
+      beforeListen: true,
+    });
     config.customFields.OrderLine.push({
       name: 'acceptBlueSubscriptionIds',
       type: 'int',
@@ -46,14 +55,18 @@ export type AcceptBluePluginOptions = Required<AcceptBluePluginOptionsInput>;
   compatibility: '>=2.2.0',
 })
 export class AcceptBluePlugin {
-  static options: AcceptBluePluginOptions = {
-    subscriptionStrategy: new DefaultSubscriptionStrategy(),
-  };
+  static options: AcceptBluePluginOptions;
 
   static init(options: AcceptBluePluginOptionsInput): AcceptBluePlugin {
+    let vendureHost = options.vendureHost;
+    // Strip trailing slash
+    if (vendureHost.endsWith('/')) {
+      vendureHost = vendureHost.slice(0, vendureHost.length - 1);
+    }
     this.options = {
-      ...this.options,
-      ...options,
+      subscriptionStrategy:
+        options.subscriptionStrategy ?? new DefaultSubscriptionStrategy(),
+      vendureHost,
     };
     return AcceptBluePlugin;
   }