+++ packages/vendure-plugin-accept-blue/src/api/accept-blue-handler.ts
@@ -44,13 +44,24 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
       label: [
         { languageCode: LanguageCode.en, value: 'Hosted tokenization key' },
       ],
-      ui: { component: 'password-form-input' },
     },
     pin: {
       type: 'string',
       required: false,
       label: [{ languageCode: LanguageCode.en, value: 'PIN' }],
       ui: { component: 'password-form-input' },
+    },
+    webhookSecret: {
+      type: 'string',
+      required: false,
+      label: [{ languageCode: LanguageCode.en, value: 'Webhook secret' }],
+      description: [
+        {
+          languageCode: LanguageCode.en,
+          value:
+            'Automatically filled when webhooks are created on payment method creation or update.',
+        },
+      ],
     },
     testMode: {
       type: 'boolean',
@@ -73,8 +84,11 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
     metadata
   ): Promise<CreatePaymentResult> {
     if (
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
       !isNoncePaymentMethod(metadata as any) &&
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
       !isCheckPaymentMethod(metadata as any) &&
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
       !isSavedPaymentMethod(metadata as any)
     ) {
       throw new UserInputError(`You either need to provide nonce input, check input or a saved payment method ID.
@@ -88,11 +102,13 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
       | NoncePaymentMethodInput
       | SavedPaymentMethodInput;
     const client = new AcceptBlueClient(args.apiKey, args.pin, args.testMode);
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
     const result = await service.handlePaymentForOrder(
       ctx,
       order,
       amount,
       client,
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
       input
     );
     const chargeTransactionId = result.chargeResult?.transaction?.id;
@@ -115,14 +131,7 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
     };
   },
 
-  async createRefund(
-    ctx,
-    input,
-    amount,
-    order,
-    payment,
-    args
-  ): Promise<CreateRefundResult> {
+  createRefund(): Promise<CreateRefundResult> {
     throw Error(`not implemented`);
   },
 });